FeedBack

+++ src/pages/Navbar.jsx
@@ -1,8 +1,28 @@
+import Swal from "sweetalert2";
+
 const Admin = () => {
     const logOut = () => {
-        window.localStorage.clear();
-        window.sessionStorage.clear();
+        const swalWithBootstrapButtons = Swal.mixin({
+            buttonsStyling: true
+        });
+        swalWithBootstrapButtons.fire({
+            title: "Apakah Kamu yakin?",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#28a745",
+            cancelButtonColor: "#DC3545",
+            confirmButtonText: "Yes, Log Out!",
+            cancelButtonText: "No, cancel!",
+            reverseButtons: true
+        }).then((result) => {
+            if (result.isConfirmed) {
+                window.localStorage.clear();
+                window.sessionStorage.clear();
+                window.location.href = "/";
+            }
+        });
     };
+
     return (
         <nav className="border-gray-200 bg-gray-50 dark:bg-gray-800 dark:border-gray-700">
             <div className="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto p-4">
@@ -18,10 +38,7 @@ const Admin = () => {
                 <div className="hidden w-full" id="navbar-hamburger">
                     <ul className="flex flex-col font-medium mt-4 rounded-lg bg-gray-50 dark:bg-gray-800 dark:border-gray-700">
                         <li>
-                            <a href="/" className="block py-2 px-3 text-gray-900 rounded hover:bg-gray-100 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white">Home</a>
-                        </li>
-                        <li>
-                            <a href="/" onClick={logOut} className="block py-2 px-3 text-gray-900 rounded hover:bg-gray-100 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white">Log Out</a>
+                            <a onClick={logOut} className="block py-2 px-3 text-gray-900 rounded hover:bg-gray-100 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white">Log Out</a>
                         </li>
                     </ul>
                 </div>
